requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
